requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
aiohttp=3.7.4.post0=pypi_0
alabaster=0.7.12=py_0
appdirs=1.4.4=pypi_0
argh=0.26.2=pyh9f0ad1d_1002
argon2-cffi=20.1.0=py38h1e8a9f7_1
astor=0.8.1=pypi_0
astroid=2.4.2=py38h32f6830_0
async-timeout=3.0.1=pypi_0
async_generator=1.10=py_0
atomicwrites=1.4.0=pyh9f0ad1d_0
attrs=20.2.0=pyh9f0ad1d_0
autopep8=1.5.4=pyh9f0ad1d_0
babel=2.8.0=py_0
backcall=0.2.0=pyh9f0ad1d_0
backports=1.0=py_2
backports.functools_lru_cache=1.6.1=py_0
bce-python-sdk=0.8.59=pypi_0
bcrypt=3.2.0=py38h1e8a9f7_0
bleach=3.2.1=pyh9f0ad1d_0
brotlipy=0.7.0=py38h1e8a9f7_1000
ca-certificates=2021.1.19=haa95532_1
certifi=2020.12.5=py38haa95532_0
cffi=1.14.3=py38hba49e27_0
cfgv=3.2.0=pypi_0
chardet=3.0.4=py38h9bdc248_1008
click=7.1.2=pypi_0
cloudpickle=1.6.0=py_0
colorama=0.4.3=py_0
cryptography=3.1.1=py38hba49e27_0
cycler=0.10.0=py_2
decorator=4.4.2=py_0
defusedxml=0.6.0=py_0
diff-match-patch=20200713=pyh9f0ad1d_0
discord-py=1.7.2=pypi_0
distlib=0.3.1=pypi_0
docopt=0.6.2=pypi_0
docutils=0.16=py38h32f6830_1
entrypoints=0.3=py38h32f6830_1001
et-xmlfile=1.0.1=pypi_0
filelock=3.0.12=pypi_0
flake8=3.8.4=py_0
flask=1.1.2=pypi_0
flask-babel=2.0.0=pypi_0
freetype=2.10.2=hd328e21_0
fsspec=0.8.7=pyhd8ed1ab_0
future=0.18.2=py38h32f6830_1
gast=0.3.3=pypi_0
geos=3.8.0=h33f27b4_0
icu=67.1=h33f27b4_0
identify=2.2.0=pypi_0
idna=2.10=pyh9f0ad1d_0
imageio=2.9.0=pypi_0
imagesize=1.2.0=py_0
imgaug=0.4.0=pypi_0
importlib-metadata=2.0.0=py_1
importlib_metadata=2.0.0=1
intel-openmp=2020.1=216
intervaltree=3.0.2=py_0
ipykernel=5.3.4=py38h5ca1d4c_0
ipython=7.18.1=py38h1cdfbd6_0
ipython_genutils=0.2.0=py_1
ipywidgets=7.5.1=pyh9f0ad1d_1
isort=5.6.1=py38h32f6830_0
itsdangerous=1.1.0=pypi_0
jedi=0.17.1=py38h32f6830_0
jinja2=2.11.2=pyh9f0ad1d_0
joblib=1.0.1=pyhd8ed1ab_0
jpeg=9d=he774522_0
json5=0.9.5=pyh9f0ad1d_0
jsonschema=3.2.0=py38h32f6830_1
jupyter=1.0.0=py_2
jupyter_client=6.1.7=py_0
jupyter_console=6.2.0=py_0
jupyter_core=4.6.3=py38h32f6830_1
jupyterlab=2.2.8=py_0
jupyterlab_pygments=0.1.2=pyh9f0ad1d_0
jupyterlab_server=1.2.0=py_0
keyring=21.4.0=py38h32f6830_1
kiwisolver=1.2.0=py38heaebd3c_0
lazy-object-proxy=1.4.3=py38h9de7a3e_2
libblas=3.8.0=16_mkl
libcblas=3.8.0=16_mkl
libclang=10.0.1=default_hf44288c_1
liblapack=3.8.0=16_mkl
liblapacke=3.8.0=16_mkl
libopencv=4.4.0=py38_2
libpng=1.6.37=ha81a0f5_2
libsodium=1.0.17=h2fa13f4_0
libspatialindex=1.9.3=he025d50_3
libtiff=4.1.0=h885aae3_6
libwebp-base=1.1.0=hfa6e2cd_3
lmdb=1.1.1=pypi_0
lz4-c=1.9.2=h62dcd97_2
m2w64-gcc-libgfortran=5.3.0=6
m2w64-gcc-libs=5.3.0=7
m2w64-gcc-libs-core=5.3.0=7
m2w64-gmp=6.1.0=2
m2w64-libwinpthread-git=5.0.0.4634.697f757=2
markupsafe=1.1.1=py38h1e8a9f7_1
matplotlib=3.3.2=0
matplotlib-base=3.3.2=py38hfb9ee82_0
mccabe=0.6.1=py_1
mistune=0.8.4=py38h9de7a3e_1001
mkl=2020.1=216
msys2-conda-epoch=20160418=1
multidict=5.1.0=pypi_0
nbclient=0.5.0=py_0
nbconvert=6.0.7=py38h32f6830_0
nbformat=5.0.7=py_0
nest-asyncio=1.4.1=py_0
networkx=2.5=pypi_0
nodeenv=1.5.0=pypi_0
notebook=6.1.4=py38h32f6830_0
numpy=1.19.1=py38h72c728b_0
numpydoc=1.1.0=py_1
olefile=0.46=py_0
opencv=4.4.0=py38_2
opencv-python=4.2.0.32=pypi_0
openpyxl=3.0.7=pypi_0
openssl=1.1.1j=h2bbff1b_0
packaging=20.4=pyh9f0ad1d_0
paddleocr=2.0.3=pypi_0
paddlepaddle=2.0.1=pypi_0
pandas=1.2.3=py38h4c96930_0
pandoc=2.10.1=he774522_0
pandocfilters=1.4.2=py_1
paramiko=2.7.2=pyh9f0ad1d_0
parso=0.7.0=pyh9f0ad1d_0
pathtools=0.1.2=py_1
pexpect=4.8.0=py38h32f6830_1
pickleshare=0.7.5=py38h32f6830_1001
pillow=7.2.0=py38h7011068_1
pip=20.2.3=py_0
pipreqs=0.4.10=pypi_0
pluggy=0.13.1=py38h32f6830_2
pre-commit=2.11.1=pypi_0
prometheus_client=0.8.0=pyh9f0ad1d_0
prompt-toolkit=3.0.7=py_0
prompt_toolkit=3.0.7=0
protobuf=3.15.6=pypi_0
psutil=5.7.2=py38hab1e662_1
py-opencv=4.4.0=py38h23f93f0_2
pyclipper=1.2.1=pypi_0
pycodestyle=2.6.0=pyh9f0ad1d_0
pycparser=2.20=pyh9f0ad1d_2
pycryptodome=3.10.1=pypi_0
pydocstyle=5.1.1=py_0
pyflakes=2.2.0=pyh9f0ad1d_0
pygments=2.7.1=py_0
pylint=2.6.0=py38h32f6830_0
pynacl=1.4.0=py38hd4486f3_1
pyopenssl=19.1.0=py_1
pyparsing=2.4.7=pyh9f0ad1d_0
pyqt=5.12.3=py38h6538335_1
pyqt5-sip=4.19.18=pypi_0
pyqtwebengine=5.12.1=pypi_0
pyrsistent=0.17.3=py38h1e8a9f7_1
pysocks=1.7.1=py38h32f6830_1
pytesseract=0.3.7=pyhd3deb0d_0
python=3.8.6=h60c2a47_0_cpython
python-dateutil=2.8.1=py_0
python-dotenv=0.17.1=pypi_0
python-jsonrpc-server=0.4.0=pyh9f0ad1d_0
python-language-server=0.35.1=py_0
python-levenshtein=0.12.2=pypi_0
python_abi=3.8=1_cp38
pytz=2020.1=pyh9f0ad1d_0
pywavelets=1.1.1=pypi_0
pywin32=227=py38hfa6e2cd_0
pywin32-ctypes=0.2.0=py38h32f6830_1001
pywinpty=0.5.7=py38_0
pyyaml=5.3.1=py38h1e8a9f7_0
pyzmq=19.0.2=py38h77b9d75_1
qdarkstyle=2.8.1=pyh9f0ad1d_1
qt=5.12.9=hb2cf2c5_0
qtawesome=1.0.1=pyh9f0ad1d_0
qtconsole=4.7.7=pyh9f0ad1d_0
qtpy=1.9.0=py_0
requests=2.24.0=pyh9f0ad1d_0
rope=0.18.0=pyh9f0ad1d_0
rtree=0.9.4=py38h7ad75cc_1
scikit-image=0.17.2=pypi_0
scikit-learn=0.24.1=py38ha09990b_0
scipy=1.6.0=py38h5f893b4_0
send2trash=1.5.0=py_0
setuptools=49.6.0=py38h9bdc248_2
shapely=1.7.1=pypi_0
shellcheck-py=0.7.1.1=pypi_0
six=1.15.0=pyh9f0ad1d_0
snowballstemmer=2.0.0=py_0
sortedcontainers=2.2.2=pyh9f0ad1d_0
sphinx=3.2.1=py_0
sphinxcontrib-applehelp=1.0.2=py_0
sphinxcontrib-devhelp=1.0.2=py_0
sphinxcontrib-htmlhelp=1.0.3=py_0
sphinxcontrib-jsmath=1.0.1=py_0
sphinxcontrib-qthelp=1.0.3=py_0
sphinxcontrib-serializinghtml=1.1.4=py_0
spyder=4.1.5=py38h32f6830_0
spyder-kernels=1.9.4=py38h32f6830_0
sqlite=3.33.0=he774522_1
terminado=0.9.1=py38h32f6830_0
testpath=0.4.4=py_0
threadpoolctl=2.1.0=pyh5ca1d4c_0
tifffile=2021.3.17=pypi_0
tk=8.6.10=he774522_1
toml=0.10.1=pyh9f0ad1d_0
tornado=6.0.4=py38hfa6e2cd_0
tqdm=4.59.0=pypi_0
traitlets=5.0.4=py_1
typing-extensions=3.10.0.0=pypi_0
ujson=4.0.0=py38h7ae7562_0
urllib3=1.25.10=py_0
vc=14.1=h869be7e_1
virtualenv=20.4.3=pypi_0
visualdl=2.1.1=pypi_0
vs2015_runtime=14.16.27012=h30e32a0_2
watchdog=0.10.3=py38h32f6830_1
wcwidth=0.2.5=pyh9f0ad1d_2
webencodings=0.5.1=py_1
werkzeug=1.0.1=pypi_0
wheel=0.35.1=pyh9f0ad1d_0
widgetsnbextension=3.5.1=py38h32f6830_1
win_inet_pton=1.1.0=py38h32f6830_1
wincertstore=0.2=py38h32f6830_1005
winpty=0.4.3=4
wrapt=1.11.2=py38h9de7a3e_0
xlrd=2.0.1=pypi_0
xz=5.2.5=h62dcd97_1
yaml=0.2.5=he774522_0
yapf=0.30.0=pyh9f0ad1d_0
yarg=0.1.9=pypi_0
yarl=1.6.3=pypi_0
zeromq=4.3.2=ha925a31_3
zipp=3.3.0=py_0
zlib=1.2.11=h62dcd97_1009
zstd=1.4.5=h1f3a1b7_2
